% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_diagnostics.R
\name{print.ru}
\alias{print.ru}
\title{Print method for an \code{"ru"} object}
\usage{
\method{print}{ru}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"ru"}, a result of a call to
\code{\link{ru}} or \code{\link{ru_rcpp}}.}

\item{...}{Additional arguments.  None are used in this function.}
}
\value{
The argument \code{x}, invisibly, as for all
  \code{\link[base]{print}} methods.
}
\description{
\code{print} method for class \code{"ru"}.
}
\details{
Simply prints the call to \code{ru} or \code{ru_rcpp}.
}
\seealso{
\code{\link{summary.ru}} for summaries of the simulated values
  and properties of the ratio-of-uniforms algorithm.

\code{\link{plot.ru}} for a diagnostic plot.
}
