% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ru_cpp}
\alias{ru_cpp}
\title{Generalised ratio-of-uniforms algorithm in C++}
\usage{
ru_cpp(n, d, r, a_box, l_box, u_box, logf, psi_mode, rot_mat, hscale, pars)
}
\arguments{
\item{n}{A numeric scalar. Number of simulated values required}

\item{d}{A numerical scalar.  The dimension of the target density.}

\item{r}{A numeric scalar.  Parameter of generalized ratio-of-uniforms.}

\item{a_box}{A numeric scalar. The a box parameter.}

\item{l_box}{A numeric vector. Values of biminus(r), i = 1, ...d.}

\item{u_box}{A numeric vector. Values of biplus(r), i = 1, ...d.}

\item{logf}{pointer}

\item{psi_mode}{mode}

\item{rot_mat}{rotation matrix}

\item{hscale}{scale}
}
\value{
A list containing sim_vals and sim_vals_rho.
}
\description{
Generalised ratio-of-uniforms algorithm in C++, for use by Rcpp.
}
