% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust-internal.R
\name{rust-internal}
\alias{rust-internal}
\alias{box_cox}
\alias{box_cox_vec}
\alias{optim_box_cox}
\alias{n_grid_fn}
\alias{init_psi_calc}
\title{Internal rust functions}
\usage{
box_cox(x, lambda = 1, gm = 1, lambda_tol = 1e-06)

box_cox_vec(x, lambda = 1, gm = 1, lambda_tol = 1e-06)

optim_box_cox(x, w, lambda_range = c(-3, 3), start = NULL,
  which_lam = 1:ncol(x))

n_grid_fn(d)

init_psi_calc(phi_to_psi, phi, lambda, gm, w, which_lam)
}
\description{
Internal rust functions
}
\details{
These functions are not intended to be called by the user.
}
\keyword{internal}
